refactor controllers to use a distinct index file for route configuration, and controller for controller code
