+++ src/controllers/leagues/index.ts
@@ -1,8 +1,6 @@
-import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
+import { FastifyInstance } from 'fastify';
 import * as Schemas from './schemas';
-import Container from 'typedi';
-import { LeaguesService } from '../../services/leagues';
-import { NotFoundError } from '../../lib/errors/errors';
+import * as Controllers from './controllers';
 
 export default async (fastify: FastifyInstance) => {
   fastify.get(
@@ -10,19 +8,7 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.FindById,
     },
-    async function findById(
-      request: FastifyRequest<{ Params: Schemas.IFindByIdParams }>,
-      reply: FastifyReply
-    ) {
-      const leaguesService = Container.get(LeaguesService);
-      const league = await leaguesService.findOne(request.params.id);
-
-      if (!league) {
-        reply.send(new NotFoundError(request.params.id));
-      }
-
-      reply.send(league);
-    }
+    Controllers.findById
   );
 
   fastify.post(
@@ -30,18 +16,7 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.Create,
     },
-    async function create(
-      request: FastifyRequest<{ Body: Schemas.ICreateBody }>,
-      reply: FastifyReply
-    ) {
-      const leaguesService = Container.get(LeaguesService);
-      const league = await leaguesService.create({
-        name: request.body.name,
-        ownerId: request.body.ownerId,
-      });
-
-      reply.send(league);
-    }
+    Controllers.create
   );
 
   fastify.delete(
@@ -49,15 +24,7 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.Delete,
     },
-    async function create(
-      request: FastifyRequest<{ Params: Schemas.IDeleteParams }>,
-      reply: FastifyReply
-    ) {
-      const leaguesService = Container.get(LeaguesService);
-      await leaguesService.delete(request.params.id);
-
-      reply.code(200).send('ok');
-    }
+    Controllers.deleteLeague
   );
 
   fastify.put(
@@ -65,15 +32,6 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.InviteUser
     },
-    async function inviteUser(
-      request: FastifyRequest<{ Params: Schemas.IInviteUserParams}>,
-      reply: FastifyReply
-    ) {
-      const leaguesService = Container.get(LeaguesService);
-      // TODO: Add invitedById based on JWT
-      await leaguesService.inviteUser(request.params.id, request.params.userId);
-
-      reply.code(200).send('ok');
-    }
+    Controllers.inviteUser
   );
 };
