+++ src/controllers/login/index.ts
@@ -1,9 +1,6 @@
-import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
+import { FastifyInstance } from "fastify";
 import * as Schemas from './schemas';
-import Container from "typedi";
-import { UserService } from "../../services/users";
-import sleep from "../../utils/sleep";
-import { NotFoundError, UnauthorizedError, UnexpectedErorr } from "../../lib/errors/errors";
+import * as Controllers from './controllers';
 
 export default async (fastify: FastifyInstance) => {
   fastify.post(
@@ -11,34 +8,7 @@ export default async (fastify: FastifyInstance) => {
     {
       schema:  Schemas.Login
     },
-    async function login(
-      request: FastifyRequest<{ Body: Schemas.ILogin }>,
-      reply: FastifyReply
-    ) {
-      const usersService = Container.get(UserService);
-      const user = await usersService.findOne(null, {
-        phone: request.body.phone
-      });
-
-      if (!user) {
-        await sleep(2000);
-        return reply.send(new NotFoundError('user'));
-      }
-
-      if ((await usersService.checkPassword(user.id, request.body.password)) === false) {
-        await sleep(2000);
-        return reply.send(new UnauthorizedError());
-      }
-
-      const token = await reply.jwtSign({ user: user.id }, { expiresIn: '1d' });
-      const refreshToken = await reply.jwtSign({
-        user: user.id,
-      }, { expiresIn: '90d' });
-      
-      reply
-        .setCookie('refreshToken', refreshToken)
-        .send({ token })
-    }
+    Controllers.login
   );
 
   fastify.post(
@@ -47,12 +17,7 @@ export default async (fastify: FastifyInstance) => {
       schema: Schemas.Verify,
       onRequest: [fastify.authenticate]
     },
-    async function verify(
-      _request: FastifyRequest,
-      reply: FastifyReply
-    ) {
-      reply.send('ok');
-    }
+    Controllers.verify
   )
 
   fastify.post(
@@ -60,24 +25,7 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.Refresh
     },
-    async function refresh(
-      request: FastifyRequest,
-      reply: FastifyReply
-    ) {
-      const decoded: { user: string } = await request.jwtVerify({ onlyCookie: true });
-      
-      const usersService = Container.get(UserService);
-      const user = await usersService.findOne(decoded.user);
-
-      if (!user) {
-        await sleep(2000);
-        return reply.send(new UnauthorizedError());
-      }
-
-      const token = await reply.jwtSign({ user: user.id }, { expiresIn: '1d' });
-      
-      reply.send({ token });
-    }
+    Controllers.refresh
   );
 
   fastify.post(
@@ -86,26 +34,6 @@ export default async (fastify: FastifyInstance) => {
       schema: Schemas.Logout,
       onRequest: [fastify.authenticate]
     },
-    async function logout(
-      request: FastifyRequest,
-      reply: FastifyReply
-    ) {
-      const usersService = Container.get(UserService);
-      const user = await usersService.findOne((request.user as Schemas.IAuth).user);
-
-      if (!user) {
-        await sleep(2000);
-        return reply.send(new UnauthorizedError());
-      }
-
-      try {
-        return reply
-          .clearCookie('refreshToken')
-          .status(200)
-          .send();
-      } catch (err) {
-        reply.send(new UnexpectedErorr())
-      }
-    }
+    Controllers.logout
   )
 }
