+++ src/controllers/transactions/index.ts
@@ -1,8 +1,6 @@
-import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
+import { FastifyInstance } from 'fastify';
 import * as Schemas from './schemas';
-import Container from 'typedi';
-import { NotFoundError } from '../../lib/errors/errors';
-import { TransactionsService } from '../../services/transactions';
+import * as Controllers from './controllers';
 
 export default async (fastify: FastifyInstance) => {
   fastify.get(
@@ -10,19 +8,7 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.FindById,
     },
-    async function findById(
-      request: FastifyRequest<{ Params: Schemas.IFindByIdParams }>,
-      reply: FastifyReply
-    ) {
-      const transactionsService = Container.get(TransactionsService);
-      const transaction = await transactionsService.findOne(request.params.id);
-
-      if (!transaction) {
-        reply.send(new NotFoundError('Transaction'));
-      }
-
-      reply.send(transaction);
-    }
+    Controllers.findById
   );
 
   fastify.post(
@@ -30,23 +16,6 @@ export default async (fastify: FastifyInstance) => {
     {
       schema: Schemas.Create,
     },
-    async function create(
-      request: FastifyRequest<{ Body: Schemas.ICreateBody }>,
-      reply: FastifyReply
-    ) {
-      const transactionsService = Container.get(TransactionsService);
-
-      try {
-        const transaction = await transactionsService.TRANSACTION.create({
-          amount: request.body.amount,
-          fromId: request.body.fromId,
-          toId: request.body.toId,
-        });
-
-        reply.status(200).send(transaction);
-      } catch (err) {
-        reply.send(err);
-      }
-    }
+    Controllers.create
   );
 };
